+++ activerecord/lib/active_record/connection_adapters/column.rb
@@ -150,7 +150,13 @@ module ActiveRecord
           return string unless string.is_a?(String)
           return nil if string.empty?
 
-          string_to_time "2000-01-01 #{string}"
+          dummy_time_string = "2000-01-01 #{string}"
+
+          fast_string_to_time(dummy_time_string) || begin
+            time_hash = Date._parse(dummy_time_string)
+            return nil if time_hash[:hour].nil?
+            new_time(*time_hash.values_at(:year, :mon, :mday, :hour, :min, :sec, :sec_fraction))
+          end
         end
 
         # convert something to a boolean